# resources/curation-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Curation REST API",
    "description": "The Curation REST API enables human-in-the-loop review of entity resolution decisions. Curators can browse low-confidence matches, accept or reject proposed canonical entities, assign mentions to alternative clusters, and perform bulk curation actions. The API also provides authentication, user management, audit trails, and registry/curation statistics.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/curation/decisions": {
      "get": {
        "tags": [
          "Decisions"
        ],
        "summary": "List Decisions",
        "description": "Retrieve cursor-paginated list of decisions with optional filtering.\n\nThe UI composes the four review states from the two row primitives\n(``previous_review_count`` and ``reviewed_since_placement``) and filters via\nthe two orthogonal query parameters.\n\nArgs:\n    filters: Field-level filter criteria (entity type, confidence, etc.).\n    cursor_params: Cursor-based pagination parameters.\n    user: Authenticated and verified curator.\n    service: Decision curation service (injected).\n    ever_reviewed: Filter on lifetime review existence.\n    reviewed_since_placement: Filter on review since the current placement.\n    reviewed: Deprecated alias of ``reviewed_since_placement``.",
        "operationId": "list_decisions_api_v1_curation_decisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ever_reviewed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter on whether any curator action has ever been recorded against the decision (previous_review_count > 0). Omit to disable.",
              "title": "Ever Reviewed"
            },
            "description": "Filter on whether any curator action has ever been recorded against the decision (previous_review_count > 0). Omit to disable."
          },
          {
            "name": "reviewed_since_placement",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter on whether a curator action exists since the current placement (created_at after updated_at, else created_at). Omit to disable. Combine ever_reviewed=true with reviewed_since_placement=false to list decisions that need re-visiting after an ERE update.",
              "title": "Reviewed Since Placement"
            },
            "description": "Filter on whether a curator action exists since the current placement (created_at after updated_at, else created_at). Omit to disable. Combine ever_reviewed=true with reviewed_since_placement=false to list decisions that need re-visiting after an ERE update."
          },
          {
            "name": "reviewed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deprecated alias of reviewed_since_placement. Ignored when reviewed_since_placement is provided.",
              "deprecated": true,
              "title": "Reviewed"
            },
            "description": "Deprecated alias of reviewed_since_placement. Ignored when reviewed_since_placement is provided.",
            "deprecated": true
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type",
              "title": "Entity Type"
            },
            "description": "Filter by entity type"
          },
          {
            "name": "confidence_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum confidence",
              "title": "Confidence Min"
            },
            "description": "Minimum confidence"
          },
          {
            "name": "confidence_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum confidence",
              "title": "Confidence Max"
            },
            "description": "Maximum confidence"
          },
          {
            "name": "similarity_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum similarity",
              "title": "Similarity Min"
            },
            "description": "Minimum similarity"
          },
          {
            "name": "similarity_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum similarity",
              "title": "Similarity Max"
            },
            "description": "Maximum similarity"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search text",
              "title": "Search"
            },
            "description": "Search text"
          },
          {
            "name": "ordering",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DecisionOrdering"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ordering field",
              "title": "Ordering"
            },
            "description": "Ordering field"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor from previous response",
              "title": "Cursor"
            },
            "description": "Pagination cursor from previous response"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated list of curation decisions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_DecisionSummary_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "503": {
            "description": "MongoDB unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/curation/decisions/{decision_id}/proposed-canonical-entity": {
      "get": {
        "tags": [
          "Decisions"
        ],
        "summary": "Get Proposed Canonical Entity",
        "description": "Get the proposed canonical entity for a given decision.",
        "operationId": "get_proposed_canonical_entity_api_v1_curation_decisions__decision_id__proposed_canonical_entity_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the curation decision.",
              "title": "Decision Id"
            },
            "description": "Unique identifier of the curation decision."
          }
        ],
        "responses": {
          "200": {
            "description": "The proposed canonical entity cluster for the given decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalEntityPreview"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "description": "MongoDB unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/curation/decisions/{decision_id}/alternative-canonical-entities": {
      "get": {
        "tags": [
          "Decisions"
        ],
        "summary": "Get Alternative Canonical Entities",
        "description": "Get alternative canonical entities for a given decision.",
        "operationId": "get_alternative_canonical_entities_api_v1_curation_decisions__decision_id__alternative_canonical_entities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the curation decision.",
              "title": "Decision Id"
            },
            "description": "Unique identifier of the curation decision."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of alternative canonical entity clusters for the given decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResult_CanonicalEntityPreview_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "description": "MongoDB unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/curation/decisions/{decision_id}/accept": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Accept Decision",
        "description": "Accept the proposed canonical entity match.",
        "operationId": "accept_decision_api_v1_curation_decisions__decision_id__accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the curation decision.",
              "title": "Decision Id"
            },
            "description": "Unique identifier of the curation decision."
          }
        ],
        "responses": {
          "204": {
            "description": "Decision accepted; no content returned."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/curation/decisions/{decision_id}/reject": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Reject Decision",
        "description": "Reject the proposed canonical entity match.",
        "operationId": "reject_decision_api_v1_curation_decisions__decision_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the curation decision.",
              "title": "Decision Id"
            },
            "description": "Unique identifier of the curation decision."
          }
        ],
        "responses": {
          "204": {
            "description": "Decision rejected; no content returned."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/curation/decisions/{decision_id}/assign": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Assign Decision",
        "description": "Assign the subject entity mention to a specific cluster.",
        "operationId": "assign_decision_api_v1_curation_decisions__decision_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the curation decision.",
              "title": "Decision Id"
            },
            "description": "Unique identifier of the curation decision."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Decision assigned to the specified cluster; no content returned."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/curation/decisions/bulk-accept": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Bulk Accept Decisions",
        "description": "Accept multiple decisions in a single request.",
        "operationId": "bulk_accept_decisions_api_v1_curation_decisions_bulk_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Per-decision results for the bulk accept operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/curation/decisions/bulk-reject": {
      "post": {
        "tags": [
          "Decisions"
        ],
        "summary": "Bulk Reject Decisions",
        "description": "Reject multiple decisions in a single request.",
        "operationId": "bulk_reject_decisions_api_v1_curation_decisions_bulk_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Per-decision results for the bulk reject operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanonicalEntityPreview": {
        "properties": {
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id",
            "description": "Unique identifier of the canonical entity cluster."
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "description": "Model confidence that this cluster is the correct match."
          },
          "similarity_score": {
            "type": "number",
            "title": "Similarity Score",
            "description": "Similarity score between the entity mention and the cluster."
          },
          "cluster_size": {
            "type": "integer",
            "title": "Cluster Size",
            "description": "Total number of decisions (entity mentions) assigned to this cluster, sourced from the cluster_sizes projection."
          },
          "top_entities": {
            "items": {
              "$ref": "#/components/schemas/EntityMentionPreview"
            },
            "type": "array",
            "title": "Top Entities",
            "description": "Representative entity mentions from this cluster."
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "confidence_score",
          "similarity_score",
          "cluster_size",
          "top_entities"
        ],
        "title": "CanonicalEntityPreview",
        "description": "Cluster preview with top entity mentions for display."
      },
...
      "CurationErrorResponse": {
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/CurationErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "ERS business request UUID for log/trace correlation. Populated by handlers that run after the request-id middleware."
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "CurationErrorResponse",
        "description": "Standard error response body returned by all Curation API endpoints.\n\nThe ``request_id`` field carries the ERS business UUID set by the request\nmiddleware (``set_request_id`` in ``ers.commons.adapters.tracing``). It is\npopulated only by handlers that have access to that context \u2014 primarily the\n``Exception`` (HTTP 500) handler \u2014 and is ``None`` on responses produced\nbefore the middleware ran or by handlers that do not need correlation."
      },
...
      "CursorPage_DecisionSummary_": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/DecisionSummary"
            },
            "type": "array",
            "title": "Results",
            "description": "Page of result items."
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of items returned in this page.",
            "default": 0
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor to fetch the next page, or null if there are no more pages."
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CursorPage[DecisionSummary]"
      },
...
      "DecisionOrdering": {
        "type": "string",
        "enum": [
          "confidence_score",
          "-confidence_score",
          "created_at",
          "-created_at",
          "updated_at",
          "-updated_at",
          "cluster_size",
          "-cluster_size"
        ],
        "title": "DecisionOrdering",
        "description": "Allowed ordering options for decision listing."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}